Removing Use of Grid-Based Map
Initial removing of GameMap, MapCell, ViewWindow functionality. Code
commented out but not removed.

This will serve as the base code for creating a more optimized game loop
for rendering and processing objects without the disadvantage of
multiple recursive calls when the screen needs to be redrawn. While
recursive drawing may be able to save resources if the screen is not
changing, it is not as effiicient for high FPS situations due to many
more function calls. A more specific gameloop implementation will be
developed starting with this build.