Merge Basic Gameloop with Master (#1)
* Removing Use of Grid-Based Map

Initial removing of GameMap, MapCell, ViewWindow functionality. Code
commented out but not removed.

This will serve as the base code for creating a more optimized game loop
for rendering and processing objects without the disadvantage of
multiple recursive calls when the screen needs to be redrawn. While
recursive drawing may be able to save resources if the screen is not
changing, it is not as effiicient for high FPS situations due to many
more function calls. A more specific gameloop implementation will be
developed starting with this build.

* Finished Removing Grid Map Code

Removed comments and java classes related to the grid-based map system.
Added skeletons for Sprite and GameEntity classes for the new gameloop
design to come.

* Added Sprite Rotations

POC for rotating sprite images, transitioned to using Graphics2D to
support AffineTransform for rotations. Note: Graphics2D is only needed
to support transforms drawn onto the BufferedImage of the next frame.
When the buffers switch, the current frame is drawn using a normal
graphics object.

Added missing comments at top of some source code files.

* Sprite Rotation Improvements

Some general improvements to the sprite rotation example, example image
changed to a better test image (debugship_blue is under a free to
modify, share, and use license).

Small changes to Model class, more variables moved to ModelVars class.

Further development of gameloop will be done in master.

* Compile Fix

Added shell SpaceShip class needed for previous build to compile.

Further development of the gameloop will be done in the master branch.